Don't use the type of checked TYPED-FORMs to rewrite acode
Optimizations that rely on ASSERTED-TYPE might change the value (hence type) of
the inner form if it isn't actually of the given type, which would defeat the
point of typechecking.

This looks like it may hurt the performance of code that both trusts
declarations and generates checked TYPED-FORMs, but at the time of writing, this
can only happen when the user provides a compiler policy that has both
TRUST-DECLARATIONS and DECLARATIONS-TYPECHECK enabled; in the default policy,
they're mutually exclusive.