Keep natural non-fixnum variables on the unboxed stack
Support for this was added together with support for unboxed floating-point
variables in 2f2e51b on x86 and cbf2406 on ARM, but the natural part stayed
disabled. We previously didn't recognize enough natural operations to make
unboxed variables a net win, but this has changed.

This change slows down code which performs many reads and writes of
type-declared natural variables without combining them with natural operations.
However, such code tends to be less common and performance-critical than
Ironclad-like code that performs lots of natural ops on natural variables.

The "correct" approach would be to check how the variable is used in the body,
but that's a lot more involved.