Punt based on rewritten initforms during acode rewriting
ACODE-MAYBE-PUNT-VAR never actually set the puntable bit, so it could only punt
the variables marked as puntable during acode creation. Rewriting the initform
before checking puntability allows us to punt variables whose initforms became
constant after constant folding.

We now try to use the declared type of variables for ASSERTED-TYPE when
rewriting initforms; this was previously only done if the variable was puntable.

This exposes an obscure bug in NX1-TAGBODY where the index variable isn't
properly marked as being setq'd, resulting in a variable that is both setq'd and
punted.